% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callmultmoments.R
\name{callmultmoments}
\alias{callmultmoments}
\title{Compute Multivariate Moment Symbolically}
\usage{
callmultmoments(moment)
}
\arguments{
\item{moment}{A vector \code{c(k1, ..., kn)} specifying the moment
\eqn{X_1^{k_1} \cdots X_n^{k_n}}.}
}
\value{
An object of class \code{'moment'}, which is a list with three
components:
\itemize{
\item \code{moment}: The input moment vector.
\item \code{representation}: A matrix containing the representation in
terms of upper-triangular matrices.
\item \code{coefficients}: The coefficients corresponding to the rows of
the representation.
}
If the sum of the exponents is odd, returns \code{-1} and prints
"Sum of powers is odd. Moment is 0."
\cr\cr
If any exponent is negative, returns \code{-2} and prints
"All components of the moment must be non-negative."
\cr\cr
If any exponent is not an integer, returns \code{-3} and prints
"All components of the moment must be integers."
}
\description{
Computes a multivariate normal moment by initializing variables, calling
\code{multmoments}, and constructing output.
}
\details{
Each row of the representation gives the exponents for a single
product of covariance terms. For example, \code{(1, 2, 0)} represents
\eqn{S_{11}^1 S_{12}^2 S_{22}^0}, where the \eqn{S_{ij}} are the covariances.
\cr\cr
The full moment is the sum of these terms multiplied by their respective
coefficients. If the sum of the exponents is odd, the moment is 0.
}
\examples{
# Compute the moment for the 4-dimensional moment c(1,2,3,4):
m.1234 <- callmultmoments(c(1, 2, 3, 4))

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{multmoments}}, and the methods \code{\link{toLatex}},
\code{\link{evaluate}}, and \code{\link{simulate}} in the \code{symmoments}
package.
}
\author{
Kem Phillips <\email{kemphillips@comcast.net}>
}
