% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.mpoly}
\alias{convert.mpoly}
\title{Convert Between mpoly and List Representations of Multivariate Polynomials}
\usage{
convert.mpoly(poly)
}
\arguments{
\item{poly}{An \code{mpoly} object, or a list containing powers and
coefficients that define a multivariate polynomial.}
}
\value{
If \code{poly} is of class \code{'mpoly'}, returns a list with two
components (\code{powers} and \code{coeff}). If \code{poly} is a list of this
form, returns the corresponding \code{mpoly} object.
}
\description{
Converts an \code{mpoly} object to a simple list representation, or converts
a simple list representation back to an \code{mpoly} object.
}
\details{
The list representation consists of 2 components:
\itemize{
\item \code{powers}: A matrix where each row represents the exponents/powers
of \eqn{X} for a single term in the multivariate polynomial.
\item \code{coeff}: A numeric vector where each element is the coefficient
for the corresponding row/term in \code{powers}.
}
}
\examples{
\dontrun{
library(mpoly)

# Create an mpoly object
t0 <- mpoly::mpoly(list(
  c(coef = 3, x1 = 2),
  c(coef = 2, x1 = 1, x2 = 3),
  c(coef = -4, z = 2),
  c(coef = 1, x1 = 1, x2 = 2, z = 1)
))  

# Convert from mpoly to list representation
t1 <- convert.mpoly(t0)    

# Convert from list representation back to an mpoly object
t2 <- convert.mpoly(t1) 
}

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{convert.multipol}},
\code{\link{evaluate_expected.polynomial}},
\code{\link{integrate.polynomial}}
}
\author{
Kem Phillips \if{html}{\out{<\email{kemphillips@comcast.net}>}}
}
