% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert.multipol}
\alias{convert.multipol}
\title{Convert Between multipol and List Representations of Multivariate Polynomials}
\usage{
convert.multipol(poly)
}
\arguments{
\item{poly}{A \code{multipol} object, or a list containing powers and
coefficients that define a multivariate polynomial.}
}
\value{
If \code{poly} is of class \code{'multipol'}, returns a list with two
components (\code{powers} and \code{coeff}). If \code{poly} is a list of this
form, returns the corresponding \code{multipol} object.
}
\description{
Converts a \code{multipol} object to a simple list representation, or converts
a simple list representation back to a \code{multipol} object.
}
\details{
The list representation consists of 2 components:
\itemize{
\item \code{powers}: A matrix where each row represents the exponents/powers
of \eqn{X} for a single term in the multivariate polynomial.
\item \code{coeff}: A numeric vector where each element is the coefficient
for the corresponding row/term in \code{powers}.
}
}
\examples{
\dontrun{
library(mpoly)
library(multipol)

# Create an mpoly object to work with
t0 <- mpoly::mpoly(list(
  c(coef = 3, x1 = 2),
  c(coef = 2, x1 = 1, x2 = 3),
  c(coef = -4, z = 2),
  c(coef = 1, x1 = 1, x2 = 2, z = 1)
)) 

# Convert from mpoly to list representation
t1 <- convert.mpoly(t0)    

# Convert from list representation to a multipol object
t2 <- convert.multipol(t1) 

# Convert back to a list representation
t3 <- convert.multipol(t2) 
}

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{convert.mpoly}},
\code{\link{evaluate_expected.polynomial}},
\code{\link{integrate.polynomial}}
}
\author{
Kem Phillips <\email{kemphillips@comcast.net}>
}
