% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate.moment}
\alias{evaluate.moment}
\title{Evaluate a Multivariate Moment}
\usage{
\method{evaluate}{moment}(object, sigma)
}
\arguments{
\item{object}{An object of class \code{'moment'}.}

\item{sigma}{An upper-triangular matrix of covariance terms expressed as a
vector at which the moment is to be evaluated.}
}
\value{
The numeric value of the moment evaluated at the specified
covariance matrix.
}
\description{
Generic method for class \code{moment} to compute the numerical value of a
moment at a specified covariance matrix from the output of
\code{\link{callmultmoments}}.
}
\details{
\code{object} is normally the output of a call to
\code{\link{callmultmoments}}. This is a list with the first component being
the moment itself, the second component being the set of upper-triangular
matrices representing the moment, and the third component containing their
corresponding coefficients. This is an object of class \code{'moment'}.
}
\examples{
# Evaluates the moment at c(1,2,3,4) at the following covariance matrix:
#   4 2 1 1
#   2 3 1 1
#   1 1 2 1
#   1 1 1 2
evaluate(callmultmoments(c(1, 2, 3, 4)), c(4, 2, 1, 1, 3, 1, 1, 2, 1, 2))

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{callmultmoments}}, and the methods \code{\link{simulate}}
and \code{\link{toLatex}} from the \code{symmoments} package.
}
\author{
Kem Phillips <\email{kemphillips@comcast.net}>
}
