% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate a Moment or Polynomial}
\usage{
evaluate(object, sigma)
}
\arguments{
\item{object}{An object of class \code{symmoment} or a multi-dimensional polynomial.}

\item{sigma}{A numeric matrix representing the covariance matrix.}
}
\value{
The evaluated result.
}
\description{
Evaluate a Moment or Polynomial
}
