% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_expected.polynomial}
\alias{evaluate_expected.polynomial}
\title{Evaluate the Expected Value of a Multivariate Polynomial}
\usage{
evaluate_expected.polynomial(poly, mu, sigma, envir = "symmoments")
}
\arguments{
\item{poly}{An object of class \code{'mpoly'} or \code{'multipol'}, or a simple
list containing coefficients and powers defining the polynomial.}

\item{mu}{A vector of real numbers representing the mean vector \eqn{\mu}
of the multivariate normal distribution.}

\item{sigma}{A vector giving an upper-triangular matrix, stacked by row,
representing the covariance matrix of the multivariate distribution.}

\item{envir}{A character string specifying the environment containing the
central moments needed for the calculation. Defaults to \code{'symmoments'}.}
}
\value{
The expected value of the multivariate polynomial evaluated at the
specified multivariate normal mean and covariance matrix.
}
\description{
Evaluates the expected value of a multivariate polynomial assuming a
specified non-central multivariate normal distribution.
}
\details{
This function searches the environment specified in the \code{envir}
argument for the central moments required to complete the expected value expansion.
The default is the \code{symmoments} environment. The computation will stop
with an error message if any required central moment is missing from \code{envir}.
}
\examples{
\dontrun{
library(mpoly)

# Define an mpoly object for a multivariate polynomial and determine
# its expected value at a specified mean and covariance matrix:
# Note: All moments up to c(2,3,2) must exist in the symmoments environment. 
# Run make.all.moments(c(2,3,2)) beforehand if necessary.

t0 <- mpoly(list(
  c(coef = 3, x1 = 2),
  c(coef = 2, x1 = 1, x2 = 3),
  c(coef = -4, z = 2),
  c(coef = 1, x1 = 1, x2 = 2, z = 1)
))

evaluate_expected.polynomial(t0, c(1, 2, 3), c(1, 0, 0, 1, 0, 1))
}

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{evaluate_noncentral}}, \code{\link{make.all.moments}}
}
\author{
Kem Phillips <\email{kemphillips@comcast.net}>
}
