% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate_noncentral}
\alias{evaluate_noncentral}
\title{Evaluate a Non-Central Multivariate Moment}
\usage{
evaluate_noncentral(moment, mu, sigma, envir = "symmoments")
}
\arguments{
\item{moment}{A vector of non-negative integers representing the
non-central moment to be evaluated: \eqn{X_1^{k_1} \cdots X_n^{k_n}}.}

\item{mu}{A vector of real numbers representing the mean vector \eqn{\mu}
of the multivariate normal distribution.}

\item{sigma}{An upper-triangular matrix of covariance terms for the
multivariate normal distribution, expressed as a vector stacked by row,
at which the moment is to be evaluated.}

\item{envir}{A character string specifying the environment containing the
central moments needed for the calculation. Defaults to \code{'symmoments'}.}
}
\value{
The numeric value of the non-central moment evaluated at the
specified mean and covariance matrix.
}
\description{
Computes the numerical value of a non-central moment at a specified mean
and specified covariance matrix.
}
\details{
This function searches the environment specified in the \code{envir}
argument for the central moments required to complete the non-central expansion.
The default is the \code{symmoments} environment. All even central moments
less than or equal to the requested \code{moment} vector must be present.
The computation will stop with an error message if any required central moment
is missing from \code{envir}.
}
\examples{
\dontrun{
# Evaluates the expected value of X1^3 X2 X3^2 at mean c(3,4,1) 
# and at the following covariance matrix:
#    4 2 1 
#    2 3 1 
#    1 1 2 
# Note: requires all central moments up to c(3,1,2) to exist in 'symmoments'.
# If needed, run: make.all.moments(c(3,1,2))
evaluate_noncentral(c(3, 1, 2), c(3, 4, 1), c(4, 2, 1, 3, 1, 2))

# Using central moments stored instead in the global environment:
evaluate_noncentral(c(3, 1, 2), c(3, 4, 1), c(4, 2, 1, 3, 1, 2), '.GlobalEnv')
}

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{evaluate.moment}}, \code{\link{make.all.moments}}
}
\author{
Kem Phillips \email{kemphillips@comcast.net}
}
