% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.polynomial.R
\name{integrate.polynomial}
\alias{integrate.polynomial}
\title{Numerically Integrate a Multivariate Polynomial}
\usage{
integrate.polynomial(poly, mu, sigma, lower = NULL, upper = NULL)
}
\arguments{
\item{poly}{An object of class \code{'mpoly'} or \code{'multipol'}, or a
simple list containing two components (\code{coeff} and \code{powers})
defining the polynomial.}

\item{mu}{A numeric vector giving the mean vector \eqn{\mu} of the
multivariate normal distribution.}

\item{sigma}{A square matrix specifying the covariance matrix of the
multivariate normal distribution.}

\item{lower}{A numeric vector of the lower limits of integration, containing
one element for each dimension. If \code{NULL} (the default), it defaults
to \eqn{-6} times the standard deviations from the mean.}

\item{upper}{A numeric vector of the upper limits of integration, containing
one element for each dimension. If \code{NULL} (the default), it defaults
to \eqn{+6} times the standard deviations from the mean.}
}
\value{
The expected value of the polynomial numerically integrated against
the specified multivariate normal distribution.
}
\description{
Integrates a multivariate polynomial against a specified non-central
multivariate normal distribution using ordinary numerical integration via
the \code{\link[cubature]{adaptIntegrate}} function from the \code{cubature} package.
}
\details{
Defaults for \code{lower} and \code{upper} boundaries are set to
\eqn{\pm 6} times the standard deviations (the square roots of the diagonal
elements of the covariance matrix \code{sigma}).
\cr\cr
If the polynomial is defined by a simple list, it must contain two components:
\itemize{
\item \code{powers}: A matrix where each row represents the exponents/powers
for a single term in the polynomial.
\item \code{coeff}: A numeric vector where each element is the coefficient
of the corresponding row in \code{powers}.
}
For example, the list structure equivalent to the polynomial in the examples section is:
\cr
\code{list(coeff = c(3, 2, -4, 1), powers = matrix(c(2,0,0, 1,3,0, 0,0,2, 1,2,1), ncol = 3, byrow = TRUE))}
}
\examples{
\dontrun{
library(mpoly)

# Define an mpoly object for a multivariate polynomial
t0 <- mpoly(list(
  c(coef = 3, x1 = 2),
  c(coef = 2, x1 = 1, x2 = 3),
  c(coef = -4, z = 2),
  c(coef = 1, x1 = 1, x2 = 2, z = 1)
))

# Numerically integrate against a specified mean and covariance identity matrix
integrate.polynomial(t0, c(1, 2, 3), matrix(c(1,0,0, 0,1,0, 0,0,1), nrow = 3, byrow = TRUE))
}

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{evaluate_expected.polynomial}}, \code{\link{multmoments}},
\code{\link{evaluate}}, \code{\link{simulate}}
}
\author{
Kem Phillips \email{kemphillips@comcast.net}
}
