% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.all.moments.R
\name{make.all.moments}
\alias{make.all.moments}
\title{Create All Moments Up to Specified Size in Environment symmoments}
\usage{
make.all.moments(moment, verbose = TRUE)
}
\arguments{
\item{moment}{A vector \code{c(k1, ..., kn)} specifying the highest moment
to compute.}

\item{verbose}{If \code{TRUE} (the default), the names of the moments are
shown as the algorithm progresses; if \code{FALSE}, progress is hidden.}
}
\value{
All objects of class \code{'moment'} up to the value given in
\code{moment} are created in the environment \code{symmoments}.
}
\description{
Create all central moment objects of a specified or smaller size in the
\code{symmoments} environment.
}
\details{
Unsorted moments (those whose exponents are not in numeric order)
are created in the \code{symmoments} environment using the
\code{\link{tounsorted}} function to transform them from their sorted counterpart.
If the \code{symmoments} environment does not exist, the user is prompted to
create it using \code{symmoments <- new.env()}.
\cr\cr
If a sorted moment does not exist, it is automatically created.
Moments of lower dimension are not created; for example, if \code{c(2, 4)} is
input, \code{m20} is created, but \code{m2} is not.
\cr\cr
\strong{Naming Conventions:}
\itemize{
\item Moments are named using the structure \code{mij..l}, e.g., \code{m136}.
\item If any exponent is greater than 9, lowercase letters, and then
uppercase letters are used. For example, \code{m3bA} represents the
moment \code{c(3, 11, 36)}.
\item The largest single exponent allowed by this alphanumeric encoding
scheme is \eqn{9 + 26 + 26 = 61}.
}
If an object with a name of this form already exists in the target environment
but is not an object of class \code{"moment"}, it will be silently overwritten
by the new moment object.
}
\examples{
\dontrun{
# Create the symmoments environment if it does not exist 
symmoments <- new.env()

# Compute all moments up to c(3,3)
make.all.moments(c(3, 3))
}

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{callmultmoments}}, \code{\link{tounsorted}}
}
\author{
Kem Phillips \email{kemphillips@comcast.net}
}
