% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multmoments.R
\name{multmoments}
\alias{multmoments}
\title{Recursive Function to Compute a Multivariate Moment}
\usage{
multmoments(moment, current.matrix, current.cell, moment.rep, row_col)
}
\arguments{
\item{moment}{A vector \code{c(k1, ..., kn)} specifying the moment
\eqn{X_1^{k_1} \cdots X_n^{k_n}}.}

\item{current.matrix}{Upper-triangular integer matrix under consideration
in the recursion.}

\item{current.cell}{Cell in the current matrix under consideration in
the recursion.}

\item{moment.rep}{Current set of representations; \code{multmoments} adds
each satisfying matrix to \code{moment.rep}.}

\item{row_col}{Matrix giving rows and columns for a square matrix for
each cell.}
}
\value{
The moment representation, \code{moment.rep}, augmented with
additional representations.
}
\description{
Called by \code{\link{callmultmoments}} to compute the representation of a
multivariate normal moment using a recursive algorithm.
}
\details{
Each row of the representation gives the exponents for a single
product of covariance terms. For example, \code{(1, 2, 0)} represents
\eqn{S_{11}^1 S_{12}^2 S_{22}^0}, where the \eqn{S_{ij}} are the covariances.
\cr\cr
This function would normally only be called by \code{\link{callmultmoments}}.
}
\references{
Phillips, K. (2010). Symbolic Computation of the Central
Moments of the Multivariate Normal Distribution. \emph{Journal of
Statistical Software}.
}
\seealso{
\code{\link{callmultmoments}}
}
\author{
Kem Phillips \if{html}{\out{<\email{kemphillips@comcast.net}>}}
}
\keyword{internal}
