% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.moment.R
\name{print.moment}
\alias{print.moment}
\title{Print the Representation of a Multivariate Moment}
\usage{
\method{print}{moment}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'moment'}, usually the output of
\code{\link{callmultmoments}}.}

\item{...}{Included only for consistency with the generic function.}
}
\description{
Prints an object of class \code{'moment'}.
}
\details{
Prints the moment as \code{E[X1**k1 X2**k2 ...]}: followed by
the lines of the representation matrix with the corresponding coefficient
attached to each row.
}
\examples{
print(callmultmoments(c(1, 2, 3)))

}
\references{
Phillips, K. (2010). Symbolic Computation of the Central
Moments of the Multivariate Normal Distribution. \emph{Journal of
Statistical Software}.
}
\seealso{
\code{\link{callmultmoments}}
}
\author{
Kem Phillips <\email{kemphillips@comcast.net}>
}
