% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.moment.R
\name{simulate.moment}
\alias{simulate.moment}
\title{Compute a Multivariate Moment Using Monte Carlo Integration}
\usage{
\method{simulate}{moment}(object, nsim, seed = NULL, Mean, Sigma, ...)
}
\arguments{
\item{object}{An object of class \code{'moment'} representing
\eqn{E[X_1^{k_1} \cdots X_n^{k_n}]}.}

\item{nsim}{The number of samples to generate in computing the integral.}

\item{seed}{An integer for the random number generator (\code{\link[base]{set.seed}}).}

\item{Mean}{The mean vector of \eqn{(X_1, \dots, X_n)}.}

\item{Sigma}{Covariance matrix of \eqn{(X_1, \dots, X_n)}, dimension \eqn{n \times n},
expressed as a vector stacked by row.}

\item{...}{Included only for consistency with the generic function.}
}
\value{
An approximate numerical value of the specified moment.
}
\description{
Computes a multivariate normal moment by Monte Carlo integration.
}
\note{
Non-central moments can be approximated by specifying \code{Mean}.
For central moments, set \code{Mean} to a vector of zeros.
\cr\cr
The \code{mvtnorm} package must be installed for this function to utilize
\code{\link[mvtnorm]{rmvnorm}}.
}
\examples{
# Using 10000 samples, estimate the central moment for the moment c(2,4) 
# at the specified covariance matrix and mean (0,0):
library(mvtnorm)
simulate(callmultmoments(c(2, 4)), 10000, NULL, c(0, 0), c(2, 1, 1, 4))

}
\references{
\insertRef{Rizzo2008}{symmoments}
}
\seealso{
\code{\link{callmultmoments}}, and the methods \code{\link{toLatex}}
and \code{\link{evaluate}}.
}
\author{
Kem Phillips <\email{kemphillips@comcast.net}>
}
