% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatex.R
\name{toLatex.moment}
\alias{toLatex.moment}
\title{LaTeX a Multivariate Moment}
\usage{
\method{toLatex}{moment}(object, ...)
}
\arguments{
\item{object}{An object of class \code{'moment'}, usually the output of
\code{\link{callmultmoments}}.}

\item{...}{Included only for consistency with the generic function.}
}
\value{
A character vector giving the LaTeX code for the symbolic moment.
}
\description{
Computes a LaTeX representation sorted lexicographically of an object of
class \code{'moment'}.
}
\details{
The first element of the result is the moment expressed as an
expected value (\code{E[...] =}). The remaining lines are the LaTeX
representation broken at appropriate intervals for printing. (Individual
terms for high dimensions will still overrun a printed line.)
\cr\cr
Double backslashes (\code{\\\\}) are inserted where LaTeX requires a
backslash. These can be reset to single backslashes by writing the output
to a file using the standard R function \code{\link[base]{writeLines}}.
}
\examples{
toLatex(callmultmoments(c(1, 2, 3)))

}
\references{
\insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{callmultmoments}}, and the \code{\link{evaluate}} method.
}
\author{
Kem Phillips \email{kemphillips@comcast.net}
}
