% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLatex.R
\name{toLatex_noncentral}
\alias{toLatex_noncentral}
\title{Compute a LaTeX Expression for a Non-Central Moment}
\usage{
toLatex_noncentral(moment, envir = "symmoments")
}
\arguments{
\item{moment}{A vector \code{c(k1, ..., kn)} specifying the moment
\eqn{X_1^{k_1} \cdots X_n^{k_n}}.}

\item{envir}{A character string specifying the environment that contains
the required central moments. Defaults to \code{'symmoments'}.}
}
\value{
A character string giving the LaTeX representation of the non-central
moment where \eqn{X} follows a multivariate normal distribution.
}
\description{
Computes a LaTeX expression for a non-central multivariate normal moment.
}
\details{
All required central moment objects must already exist in the
specified environment (the default is \code{'symmoments'}). However, if only
the sorted version of an unsorted moment exists in that environment, the
\code{\link{tounsorted}} function will automatically be called to transform
and obtain it.
}
\examples{
\dontrun{
# Compute the LaTeX representation of the 2-dimensional non-central moment c(1,3).
# Note: This requires that all central moments up to c(1,3) have already been 
# generated in the symmoments environment using make.all.moments.
toLatex_noncentral(c(1, 3))
}

}
\references{
Phillips, K. (2010). Symbolic Computation of the Central
Moments of the Multivariate Normal Distribution. \emph{Journal of
Statistical Software}.
}
\seealso{
\code{\link{make.all.moments}}, \code{\link{tounsorted}},
\code{\link{callmultmoments}}, and the method \code{\link{toLatex}}.
}
\author{
Kem Phillips \if{html}{\out{<\email{kemphillips@comcast.net}>}}
}
