% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMatching.R
\name{toMatching}
\alias{toMatching}
\title{Convert a Phylogenetic Tree from a Moment L-Matrix to Matching Form}
\usage{
toMatching(L, type = NULL, tip.label = NULL)
}
\arguments{
\item{L}{The input tree structure. This can be an \code{L-matrix} object, a
square \eqn{L} matrix, or an \eqn{L} matrix in reduced upper-triangular (vector) form.}

\item{type}{A character string, either \code{'square'} or \code{'ut'}. This must be
specified if \code{L} is a raw matrix or vector rather than a formal \code{L-matrix} object.
Defaults to \code{NULL}.}

\item{tip.label}{A character vector containing custom labels for the tips. If \code{NULL}
(the default), labels fallback to \code{"a"} through \code{"z"} if there are at most 26 tips;
otherwise, 3-letter combinations of the form \code{"aaa"}, \code{"aab"}, etc., are generated.}
}
\value{
A matching representation of the phylogenetic tree corresponding to the input.
The output list is assigned the class \code{'L-matching'}, which contains 5 components
including the tree in matching format.
}
\description{
Converts a tree structure represented in a moment format into an \code{ape}
matching format structure.
}
\details{
An \code{L-matrix} object is a list containing the following 5 components:
\itemize{
\item \code{L}: The L-matrix in full square form.
\item \code{L.ut}: The L-matrix in reduced upper-triangular form.
\item \code{Newick}: The Newick string representation of the tree structure.
\item \code{tip.label}: A character vector of the tip labels.
\item \code{tip.label.n}: An integer specifying the total number of tips.
}
}
\examples{
# Create a Newick character string
exam.Newick <- "(((a,b),c),d);"

# Convert to a moment L-matrix
exam.moment <- toMoment(exam.Newick)

# Convert to matching format
exam.matching <- toMatching(exam.moment)

}
\references{
Phillips, K. (2010). Symbolic Computation of the Central Moments of the
Multivariate Normal Distribution. \emph{Journal of Statistical Software}.
\cr\cr
Diaconis, P. W., & Holmes, S. (1998). Matchings and Phylogenetic Trees.
\emph{Proceedings of the National Academy of Sciences}, 95(25), 14600-14602.
}
\seealso{
\code{\link{toMoment}}, \code{\link{toNewick}}
}
\author{
Kem Phillips <\email{kemphillips@comcast.net}>
}
