% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMoment.R
\name{toMoment}
\alias{toMoment}
\title{Convert a Tree from Newick or Matching to Moment Format}
\usage{
toMoment(inputobject, tip.label = NULL)
}
\arguments{
\item{inputobject}{A tree structure represented as a Newick format character
string, or a \code{matching} object as defined in the \pkg{ape} package.}

\item{tip.label}{A character vector specifying rearranged labels for the
tips. If provided, these must be the original tip labels. Defaults to \code{NULL}.}
}
\value{
A moment L-matrix object corresponding to the input phylogenetic
tree object.
}
\description{
Converts a phylogenetic tree from a Newick character string or an \code{ape}
matching matrix into a moment L-matrix object.
}
\details{
The returned L-matrix class object consists of 5 internal components:
\itemize{
\item \code{L}: The L-matrix represented in full square form.
\item \code{L.ut}: The L-matrix represented in upper-triangular form.
\item \code{Newick}: The Newick string representation of the tree structure.
\item \code{tip.label}: A character vector holding the labels of the tips.
\item \code{tip.label.n}: An integer specifying the total number of tips.
}
}
\examples{
# Create a Newick character string
exam.Newick <- "(((a,b),c),d);"

# Convert to a moment L-matrix
exam.moment <- toMoment(exam.Newick)

# Convert to a matching object
exam.matching <- toMatching(exam.moment)

# Convert back to a moment object
backto.moment <- toMoment(exam.matching)

}
\references{
\insertRef{Phillips2010}{symmoments}
\cr\cr
Felsenstein, J. (1990). The Newick tree format.
\url{http://evolution.genetics.washington.edu/phylip/newicktree.html}
\cr\cr
\insertRef{Diaconis1998}{symmoments}
}
\seealso{
\code{\link{toNewick}}, \code{\link{toMatching}}
}
\author{
Kem Phillips \email{kemphillips@comcast.net}
}
