% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNewick.R
\name{toNewick}
\alias{toNewick}
\title{Convert a Phylogenetic Tree from a Moment L-Matrix to Newick Form}
\usage{
toNewick(L, type = NULL, tip.label = NULL)
}
\arguments{
\item{L}{The input tree structure. This can be an \code{L-matrix} object, a
square \eqn{L} matrix, or an \eqn{L} matrix in reduced upper-triangular (vector) form.}

\item{type}{A character string, either \code{'square'} or \code{'ut'}. This must be
specified if \code{L} is a raw matrix or vector rather than a formal \code{L-matrix} object.
Defaults to \code{NULL}.}

\item{tip.label}{A character vector containing custom labels for the tips. If \code{NULL}
(the default), labels fallback to \code{"a"} through \code{"z"} if there are at most 26 tips;
otherwise, 3-letter combinations of the form \code{"aaa"}, \code{"aab"}, etc., are generated.}
}
\value{
A character string representing the Newick format of the phylogenetic tree corresponding
to the input. The output list is assigned the class \code{'L-Newick'}, which contains 5 components
including the tree string.
}
\description{
Converts a tree structure represented in a moment format back into a Newick format string.
}
\details{
An \code{L-matrix} object is a list containing the following 5 components:
\itemize{
\item \code{L}: The L-matrix in full square form.
\item \code{L.ut}: The L-matrix in reduced upper-triangular form.
\item \code{Newick}: The Newick string representation of the tree structure.
\item \code{tip.label}: A character vector of the tip labels.
\item \code{tip.label.n}: An integer specifying the total number of tips.
}
}
\examples{
# Create a Newick character string
exam.Newick <- "(((a,b),c),d);"

# Convert to a moment L-matrix
exam.moment <- toMoment(exam.Newick)

# Convert back to Newick format
backto.Newick <- toNewick(exam.moment)

}
\references{
\insertRef{Phillips2010}{symmoments}
\cr\cr
Felsenstein, J. (1990). The Newick tree format.
\url{http://evolution.genetics.washington.edu/phylip/newicktree.html}
\cr\cr
\insertRef{Diaconis1998}{symmoments}
}
\seealso{
\code{\link{toMoment}}, \code{\link{toMatching}}
}
\author{
Kem Phillips \email{kemphillips@comcast.net}
}
