% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tounsorted.R
\name{tounsorted}
\alias{tounsorted}
\title{Compute an Unsorted Central Moment Object from a Sorted Object}
\usage{
tounsorted(moment, sorted.moment)
}
\arguments{
\item{moment}{The unsorted target moment to obtain, specified in vector
form (e.g., \code{c(3, 1, 2)}).}

\item{sorted.moment}{A sorted object of class \code{'moment'} to use as the
base for creating the unsorted moment.}
}
\value{
An object of class \code{'moment'}, which is a list containing the
following three components:
\item{moment}{The input unsorted moment vector.}
\item{representation}{A matrix containing the representation in terms of
upper-triangular matrices, rearranged to match the target unsorted order.}
\item{coefficients}{A numeric vector of the coefficients corresponding to
the rows of the representation matrix.}
}
\description{
Produces an unsorted central moment object from a sorted object of class
\code{'moment'}.
}
\details{
Unsorted moments are those whose exponents are not in sorted
numerical order (e.g., \code{m312} vs \code{m123}). The unsorted moment's
representation is calculated by rearranging the rows and columns of the
sorted moment's matrices successively.
}
\examples{
# Obtain unsorted moment m312 from sorted base m123
tounsorted(c(3, 1, 2), callmultmoments(c(1, 2, 3)))

}
\references{
#' \insertRef{Phillips2010}{symmoments}
}
\seealso{
\code{\link{multmoments}}, \code{\link{callmultmoments}}
}
\author{
Kem Phillips <\email{kemphillips@comcast.net}>
}
